requirements.txt
# File: requirements.txt
# Purpose: Python dependencies for the Trading Journal application

# Web Framework
fastapi>=0.115.0,<1.0.0
uvicorn>=0.27.1,<1.0.0

# Database
sqlalchemy==2.0.27
alembic==1.13.1
pydantic>=2.8.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0

# Firebase
firebase-admin==6.2.0
google-cloud-firestore==2.11.1

# Security
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
python-multipart==0.0.9
email-validator==2.1.0

# Utilities
requests==2.31.0
python-dateutil==2.8.2
aiohttp==3.9.3
aiofiles==23.2.1
pillow==10.2.0

# Data Science
pandas==2.2.0
numpy==1.26.3
scipy==1.12.0

# MCP (Model Context Protocol) & AI
# Using official Anthropic MCP packages
mcp==1.1.0
anthropic>=0.40.0,<1.0.0

# Testing
pytest==7.4.0
pytest-asyncio==0.23.4
httpx>=0.27.0,<1.0.0

# Development
black==24.1.1
isort==5.13.2
flake8==7.0.0